Move ebnf ast to its own package
