+++ internal/ebnf/parser/ast/ast_test.go
@@ -1,4 +1,4 @@
-package parser
+package ast
 
 import (
 	"testing"
